refactor(analysis): rename parserdialog for all parsers (#4008)
