+++ analysis/analysers/parsers/GitLogParser/src/main/kotlin/de/maibornwolff/codecharta/parser/gitlogparser/GitLogParser.kt
@@ -164,7 +164,7 @@ class GitLogParser(
         )
     }
 
-    override fun getDialog(): ParserDialogInterface = ParserDialog
+    override fun getDialog(): ParserDialogInterface = Dialog
 
     override fun isApplicable(resourceToBeParsed: String): Boolean {
         println("Checking if GitLogParser is applicable...")
