+++ analysis/analysers/parsers/RawTextParser/src/main/kotlin/de/maibornwolff/codecharta/parser/rawtextparser/Dialog.kt
@@ -9,7 +9,7 @@ import de.maibornwolff.codecharta.dialogProvider.myPromptDefaultFileFolderInput
 import de.maibornwolff.codecharta.dialogProvider.myPromptInput
 import de.maibornwolff.codecharta.dialogProvider.myPromptInputNumber
 
-class ParserDialog {
+class Dialog {
     companion object : ParserDialogInterface {
         override fun collectParserArgs(session: Session): List<String> {
             val inputFileName = session.myPromptDefaultFileFolderInput(
