+++ analysis/analysers/parsers/SourceCodeParser/src/main/kotlin/de/maibornwolff/codecharta/parser/sourcecodeparser/Dialog.kt
@@ -9,7 +9,7 @@ import de.maibornwolff.codecharta.dialogProvider.myPromptDefaultFileFolderInput
 import de.maibornwolff.codecharta.dialogProvider.myPromptInput
 import de.maibornwolff.codecharta.dialogProvider.myPromptList
 
-class ParserDialog {
+class Dialog {
     companion object : ParserDialogInterface {
         override fun collectParserArgs(session: Session): List<String> {
             val inputFileName: String = session.myPromptDefaultFileFolderInput(
