+++ analysis/analysers/parsers/SourceCodeParser/src/main/kotlin/de/maibornwolff/codecharta/parser/sourcecodeparser/SourceCodeParserMain.kt
@@ -177,7 +177,7 @@ class SourceCodeParserMain(
 
     private fun getJsonOutputStream() = OutputFileHandler.stream(outputFile?.absolutePath, output, compress)
 
-    override fun getDialog(): ParserDialogInterface = ParserDialog
+    override fun getDialog(): ParserDialogInterface = Dialog
 
     override fun isApplicable(resourceToBeParsed: String): Boolean {
         println("Checking if SourceCodeParser is applicable...")
