+++ analysis/analysers/parsers/SourceCodeParser/src/test/kotlin/de/maibornwolff/codecharta/parser/sourcecodeparser/DialogTest.kt
@@ -5,7 +5,7 @@ import com.varabyte.kotter.runtime.RunScope
 import com.varabyte.kotter.runtime.terminal.inmemory.press
 import com.varabyte.kotter.runtime.terminal.inmemory.type
 import com.varabyte.kotterx.test.foundation.testSession
-import de.maibornwolff.codecharta.parser.sourcecodeparser.ParserDialog.Companion.collectParserArgs
+import de.maibornwolff.codecharta.parser.sourcecodeparser.Dialog.Companion.collectParserArgs
 import io.mockk.every
 import io.mockk.mockkObject
 import io.mockk.unmockkAll
@@ -26,7 +26,7 @@ class DialogTest {
 
     @BeforeEach
     fun setup() {
-        mockkObject(ParserDialog.Companion)
+        mockkObject(Dialog.Companion)
     }
 
     @AfterEach
@@ -72,7 +72,7 @@ class DialogTest {
                 terminal.press(Keys.ENTER)
             }
 
-            every { ParserDialog.Companion.testCallback() } returnsMany listOf(
+            every { Dialog.Companion.testCallback() } returnsMany listOf(
                 fileCallback,
                 formatCallback,
                 outFileCallback,
@@ -142,7 +142,7 @@ class DialogTest {
                 terminal.press(Keys.ENTER)
             }
 
-            every { ParserDialog.Companion.testCallback() } returnsMany listOf(
+            every { Dialog.Companion.testCallback() } returnsMany listOf(
                 fileCallback,
                 formatCallback,
                 outFileCallback,
@@ -208,7 +208,7 @@ class DialogTest {
                 terminal.press(Keys.ENTER)
             }
 
-            every { ParserDialog.Companion.testCallback() } returnsMany listOf(
+            every { Dialog.Companion.testCallback() } returnsMany listOf(
                 fileCallback,
                 formatCallback,
                 outFileCallback,
